importer - better support for kmz in upload
Previously, kmz was unpacked early in the process and alongside
any other uploaded files. This resulted in the layer/featuretype
name being 'doc' and makes tracking any related files more
difficult.

- add unpack method to DataFormat to support custom unpacking
- implement unpack for KMLFileFormat
- format test for unpack
- rest endpoint test for kmz and correct name
- support kmz named as kml (google earth does this, too)

(cherry picked from commit 2375888fbabea402302bb3c10af785d24493dc29)

Conflicts:
	src/extension/importer/core/src/test/java/org/geoserver/importer/ImporterDataTest.java

+++ src/extension/importer/core/src/main/java/org/geoserver/importer/Directory.java
@@ -88,16 +88,24 @@ public class Directory extends FileData {
     }
 
     public void unpack(File file) throws IOException {
-        //if the file is an archive, unpack it
-        VFSWorker vfs = new VFSWorker();
-        if (vfs.canHandle(file)) {
-            LOGGER.fine("unpacking " + file.getAbsolutePath() + " to " + this.file.getAbsolutePath());
-            vfs.extractTo(file, this.file);
+        DataFormat fileFormat = DataFormat.lookup(file);
+        // if unrecognized, try to unpack, otherwise skip and let the
+        // format itself do what it needs. This is important for KMZ as
+        // we want to use the kmz basename for the layer.
+        if (fileFormat == null) {
+            //if the file is an archive, unpack it
+            VFSWorker vfs = new VFSWorker();
+            if (vfs.canHandle(file)) {
+                LOGGER.fine("unpacking " + file.getAbsolutePath() + " to " + this.file.getAbsolutePath());
+                vfs.extractTo(file, this.file);
 
-            LOGGER.fine("deleting " + file.getAbsolutePath());
-            if (!file.delete()) {
-                throw new IOException("unable to delete file");
+                LOGGER.fine("deleting " + file.getAbsolutePath());
+                if (!file.delete()) {
+                    throw new IOException("unable to delete file");
+                }
             }
+        } else {
+            fileFormat.unpack(file);
         }
     }
